levelColorsをappへ移動
