+++ app/common/levelColors.ts
@@ -1,0 +1,10 @@
+export const levelColors = [
+  "text-emerald-700 dark:text-emerald-400 ",
+  "text-amber-700 dark:text-amber-400 ",
+  "text-rose-700 dark:text-rose-400 ",
+];
+export const levelBgColors = [
+  "bg-emerald-500 dark:bg-emerald-900 ",
+  "bg-amber-500 dark:bg-amber-900 ",
+  "bg-rose-400 dark:bg-rose-900 ",
+];
