+++ app/main/about/[aboutIndex]/4-level.tsx
@@ -1,4 +1,5 @@
-import { levelColors, levelTypes } from "@/../chartFormat/chart.js";
+import { levelTypes } from "@/../chartFormat/chart.js";
+import { levelColors } from "@/common/levelColors";
 
 export function AboutContent4() {
   return (
