+++ chartFormat/chart.ts
@@ -99,16 +99,6 @@ export interface Level {
   unlisted: boolean;
 }
 export const levelTypes = ["Single", "Double", "Maniac"];
-export const levelColors = [
-  "text-emerald-700 dark:text-emerald-400 ",
-  "text-amber-700 dark:text-amber-400 ",
-  "text-rose-700 dark:text-rose-400 ",
-];
-export const levelBgColors = [
-  "bg-emerald-500 dark:bg-emerald-900 ",
-  "bg-amber-500 dark:bg-amber-900 ",
-  "bg-rose-400 dark:bg-rose-900 ",
-];
 
 export const chartMaxSize = 1000000;
